requirements.txt
requests

Pillow

PyYAML
